Replace deploy command to fix queuing behavior
This was needed, because a regular `systemctl start` is a no-op if the
unit is already in progress. This meant that only the first of several
"overlapping" webhook calls would result in a deploy. The later ones
wouldn't trigger a new deploy.

The new command is a bit of a hack, but it works. It does queuing, but
coalesces the queue as well:

- If a deploy is not in progress, it will start one (#1).
- If a deploy is already in progress (#1), it will postpone the new
  request (#2) to start directly after #1 is done.
- If a deploy is queued (#2), #3 will be merged with #2. So there will
  never be a #3 (or #4, etc.) in queue.